Separate preset files & finish hell preset
