+++ soh/soh/Enhancements/Presets/Presets.cpp
@@ -1,4 +1,4 @@
-#include "presets.h"
+#include "Presets.h"
 #include <variant>
 #include <string>
 #include <cstdint>
