+++ soh/soh/Enhancements/randomizer/randomizer.cpp
@@ -17,7 +17,7 @@
 #include <imgui_internal.h>
 #include "../custom-message/CustomMessageTypes.h"
 #include "../item-tables/ItemTableManager.h"
-#include "../presets.h"
+#include "../Presets/Presets.h"
 #include "../../../src/overlays/actors/ovl_En_GirlA/z_en_girla.h"
 #include <stdexcept>
 #include "randomizer_check_objects.h"
