+++ soh/soh/OTRGlobals.cpp
@@ -43,7 +43,7 @@
 #include "window/gui/resource/Font.h"
 #include <utils/StringHelper.h>
 #include "Enhancements/custom-message/CustomMessageManager.h"
-#include "Enhancements/presets.h"
+#include "Enhancements/Presets/Presets.h"
 #include "util.h"
 #include <boost_custom/container_hash/hash_32.hpp>
